+++ tests/unitary_tests/208_test_interpolate_Npts.py
@@ -2,7 +2,8 @@
 import nrv
 import matplotlib.pyplot as plt
 import numpy as np
-from tqdm import tqdm
+from rich.progress import track
+
 
 if __name__ == "__main__":
     N_test = "208"
@@ -77,7 +78,7 @@ if __name__ == "__main__":
     bounds = [t_bound, I_bound, t_bound, I_bound, t_bound, I_bound, t_bound, I_bound, t_bound, I_bound, (0.01,0.5)]
 
 
-    for i in tqdm(range(100000)):
+    for i in track(range(100000)):
         X = np.array([np.random.uniform(low=bounds[k][0], high=bounds[k][1]) for k in range(len(bounds))])
         waveform2 = nrv.interpolate_Npts(X, plot=False, generatefigure=False, **kwrgs_interp)