+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to NRV are sumed up in this file.
 ### Added
 - improved `eval` in `_FEMResults` for serialized calls (added state variables)
 - `spec_loader.py` mostly for typing
+- progress bars are now handled with [rich.progress](https://rich.readthedocs.io/en/stable/progress.html)
 
 ### Fixed